backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic-settings==2.9.1
httpx==0.28.1
supabase==2.15.1
python-telegram-bot==22.0
gunicorn==23.0.0
